backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pandas==2.2.0
groq==0.13.0
rank-bm25==0.2.2
pydantic==2.5.3
python-multipart==0.0.6

# Production RAG pipeline
sentence-transformers==3.3.1
qdrant-client==1.12.1
pymysql==1.1.1
numpy>=1.24.0
